api/requirements.txt
langchain==1.0.3
langchain-openai==1.0.1
langgraph==1.0.2
langsmith==0.4.38
fastapi==0.120.3
uvicorn==0.38.0
langserve==0.3.3
sse-starlette==3.0.3
python-dotenv==1.2.1
requests==2.32.5
pytest==8.4.2
pytest-asyncio==1.2.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pymupdf==1.26.6
langchain-ollama==0.2.0